+++ CHANGELOG.md
@@ -3,5 +3,13 @@
 # intellij-lets Changelog
 
 ## [Unreleased]
-### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+
+- No info 
+
+## [0.0.1]
+### Added file type recognition
+
+- `lets.yaml` and `lets.*.yaml` files will be recognized as lets configs
+### Added basic completion
+
+- for now it will autocomplete only top level keywords